+++ requests_cache/policy/settings.py
@@ -25,6 +25,7 @@ class CacheSettings(RichMixin):
 
     allowable_codes: Iterable[int] = field(default=DEFAULT_STATUS_CODES)
     allowable_methods: Iterable[str] = field(default=DEFAULT_METHODS)
+    always_revalidate: bool = field(default=None)
     cache_control: bool = field(default=False)
     disabled: bool = field(default=False)
     expire_after: ExpirationTime = field(default=None)